requirements.txt
discord.py==2.5.2
PyGithub==2.6.1
requests==2.32.4
python-dotenv==1.1.1
aiohttp==3.12.14
